+++ backend/src/utils.py
@@ -1905,13 +1905,7 @@ class MockResponse:
         return self.content
 
 
-def truncate_float(number, decimals):
-    if isinstance(number, float):
-        str_number = str(number)
-        integer_part, decimal_part = str_number.split(".")
-        truncated_decimal_part = decimal_part[:decimals]
-        return float(f"{integer_part}.{truncated_decimal_part}")
-    return number
+
 
 
 def resize_b64_image(b64_image: str | str, new_size: tuple[int, int]) -> str:
