+++ backend/tests/test_speed_statistics.py
@@ -1,4 +1,4 @@
-from utils import read_transactions_from_csv, truncate_float
+from test_utils import read_transactions_from_csv, truncate_float
 
 header = {
     "Authorization": "Bearer eyJhbGciOiJSUzI1NiIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJ0ZXN0IiwiYXpwIjoiNDA3NDA4NzE4MTkyLmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwiYXVkIjoiNDA3NDA4NzE4MTkyLmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwic3ViIjoiMTEyMTAyODc3OTUzNDg0MzUyNDI3IiwiZW1haWwiOiJ0ZXN0QGdtYWlsLmNvbSIsImVtYWlsX3ZlcmlmaWVkIjp0cnVlLCJhdF9oYXNoIjoiYm54OW9WT1o4U3FJOTcyczBHYjd4dyIsIm5hbWUiOiJUZXN0IFVzZXIiLCJwaWN0dXJlIjoiIiwiZ2l2ZW5fbmFtZSI6IlRlc3QiLCJmYW1pbHlfbmFtZSI6IlVzZXIiLCJpYXQiOjE3MTM3MzQ0NjEsImV4cCI6OTk5OTk5OTk5OX0.eZYMQzcSRzkAq4Me8C6SNU3wduS7EIu_o5XGAbsDmU05GtyipQEb5iNJ1QiLg-11RbZFL3dvi8xKd3mpuw8b-5l6u8hwSpZg6wNPLY0zPX-EOwxeHLtev_2X5pUf1_IWAnso9K_knsK8CcmJoVsCyNNjlw3hrkChacJHGNzg0TTT1rh3oe6KCpbLvYlV6tUPfm5k3AMFZIT7Jntr38CZvs6gac6L_DhItJc3TNNUUHie2zgA29_r9YFlaEr_nGoSmBhIi-i0i0h34TL4JAb4qJkVM2YI2eTTv2HjEGtkx4mE5JvNQ0VxzHSJcCNOHh1gCiFD5c6rhvvxVeEqMkGGbCZKHX_vCgnIp0iE_OWyICjVTFPitQJ00fXLhyHyPb7q5J605tuK2iTHp2NCRJEXIAl9e0F_qASBBAfyL0C4FCBtvbnEMwtpoV1VWinkKgkI7JVH0AsyTugjXyAjxxsJxBTJT9qwZLxVBoaxgqNTOFfxvwstyq1VfCl3iBbpt71D"
@@ -11,7 +11,7 @@ def test_transaction_speed_min_5min(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -49,7 +49,7 @@ def test_transaction_speed_min_30min(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -94,7 +94,7 @@ def test_transaction_speed_min_1h(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -145,7 +145,7 @@ def test_transaction_speed_min_empty(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -167,7 +167,7 @@ def test_transaction_speed_hour_30min(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -205,7 +205,7 @@ def test_transaction_speed_hour_1h(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -246,7 +246,7 @@ def test_transaction_speed_hour_24h(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -309,7 +309,7 @@ def test_transaction_speed_hour_empty(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -331,7 +331,7 @@ def test_transaction_speed_day_6h(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -369,7 +369,7 @@ def test_transaction_speed_day_24h(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -407,7 +407,7 @@ def test_transaction_speed_day_1d(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -448,7 +448,7 @@ def test_transaction_speed_day_7d(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -494,7 +494,7 @@ def test_transaction_speed_day_1mo(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -563,7 +563,7 @@ def test_transaction_speed_day_empty(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -585,7 +585,7 @@ def test_transaction_speed_week_3d(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -623,7 +623,7 @@ def test_transaction_speed_week_7d(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -664,7 +664,7 @@ def test_transaction_speed_week_2mo(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -712,7 +712,7 @@ def test_transaction_speed_week_empty(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -734,7 +734,7 @@ def test_transaction_speed_month_7d(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -772,7 +772,7 @@ def test_transaction_speed_month_1mo(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -810,7 +810,7 @@ def test_transaction_speed_month_6mo(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -855,7 +855,7 @@ def test_transaction_speed_month_empty(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -877,7 +877,7 @@ def test_transaction_speed_year_2mo(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -915,7 +915,7 @@ def test_transaction_speed_year_2y(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -956,7 +956,7 @@ def test_transaction_speed_year_empty(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
