+++ backend/tests/test_tokens_and_cost_statistics.py
@@ -1,4 +1,4 @@
-from utils import read_transactions_from_csv
+from test_utils import read_transactions_from_csv
 
 header = {
     "Authorization": "Bearer eyJhbGciOiJSUzI1NiIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJ0ZXN0IiwiYXpwIjoiNDA3NDA4NzE4MTkyLmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwiYXVkIjoiNDA3NDA4NzE4MTkyLmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwic3ViIjoiMTEyMTAyODc3OTUzNDg0MzUyNDI3IiwiZW1haWwiOiJ0ZXN0QGdtYWlsLmNvbSIsImVtYWlsX3ZlcmlmaWVkIjp0cnVlLCJhdF9oYXNoIjoiYm54OW9WT1o4U3FJOTcyczBHYjd4dyIsIm5hbWUiOiJUZXN0IFVzZXIiLCJwaWN0dXJlIjoiIiwiZ2l2ZW5fbmFtZSI6IlRlc3QiLCJmYW1pbHlfbmFtZSI6IlVzZXIiLCJpYXQiOjE3MTM3MzQ0NjEsImV4cCI6OTk5OTk5OTk5OX0.eZYMQzcSRzkAq4Me8C6SNU3wduS7EIu_o5XGAbsDmU05GtyipQEb5iNJ1QiLg-11RbZFL3dvi8xKd3mpuw8b-5l6u8hwSpZg6wNPLY0zPX-EOwxeHLtev_2X5pUf1_IWAnso9K_knsK8CcmJoVsCyNNjlw3hrkChacJHGNzg0TTT1rh3oe6KCpbLvYlV6tUPfm5k3AMFZIT7Jntr38CZvs6gac6L_DhItJc3TNNUUHie2zgA29_r9YFlaEr_nGoSmBhIi-i0i0h34TL4JAb4qJkVM2YI2eTTv2HjEGtkx4mE5JvNQ0VxzHSJcCNOHh1gCiFD5c6rhvvxVeEqMkGGbCZKHX_vCgnIp0iE_OWyICjVTFPitQJ00fXLhyHyPb7q5J605tuK2iTHp2NCRJEXIAl9e0F_qASBBAfyL0C4FCBtvbnEMwtpoV1VWinkKgkI7JVH0AsyTugjXyAjxxsJxBTJT9qwZLxVBoaxgqNTOFfxvwstyq1VfCl3iBbpt71D"
@@ -11,7 +11,7 @@ def test_transaction_tokens_and_cost_min_5min(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -54,7 +54,7 @@ def test_transaction_tokens_and_cost_min_30min(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -120,7 +120,7 @@ def test_transaction_tokens_and_cost_min_1h(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -198,7 +198,7 @@ def test_transaction_tokens_and_cost_min_empty(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -220,7 +220,7 @@ def test_transaction_tokens_and_cost_hour_30min(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -272,7 +272,7 @@ def test_transaction_tokens_and_cost_hour_1h(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -327,7 +327,7 @@ def test_transaction_tokens_and_cost_hour_24h(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -429,7 +429,7 @@ def test_transaction_tokens_and_cost_hour_empty(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -451,7 +451,7 @@ def test_transaction_tokens_and_cost_day_6h(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -503,7 +503,7 @@ def test_transaction_tokens_and_cost_day_24h(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -555,7 +555,7 @@ def test_transaction_tokens_and_cost_day_1d(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -610,7 +610,7 @@ def test_transaction_tokens_and_cost_day_7d(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -678,7 +678,7 @@ def test_transaction_tokens_and_cost_day_1mo(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -792,7 +792,7 @@ def test_transaction_tokens_and_cost_day_empty(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -814,7 +814,7 @@ def test_transaction_tokens_and_cost_week_3d(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -866,7 +866,7 @@ def test_transaction_tokens_and_cost_week_7d(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -921,7 +921,7 @@ def test_transaction_tokens_and_cost_week_2mo(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -993,7 +993,7 @@ def test_transaction_tokens_and_cost_week_empty(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -1015,7 +1015,7 @@ def test_transaction_tokens_and_cost_month_7d(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -1067,7 +1067,7 @@ def test_transaction_tokens_and_cost_month_1mo(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -1119,7 +1119,7 @@ def test_transaction_tokens_and_cost_month_6mo(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -1185,7 +1185,7 @@ def test_transaction_tokens_and_cost_month_empty(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -1207,7 +1207,7 @@ def test_transaction_tokens_and_cost_year_2mo(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -1259,7 +1259,7 @@ def test_transaction_tokens_and_cost_year_2y(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
@@ -1317,7 +1317,7 @@ def test_transaction_tokens_and_cost_year_empty(client, application):
         repo = ctx["transaction_repository"]
         repo.delete_cascade(project_id="project-test")
         transactions = read_transactions_from_csv(
-            "../test_transactions_tokens_cost_speed.csv"
+            "test_transactions_tokens_cost_speed.csv"
         )
         for transaction in transactions:
             repo.add(transaction)
